requirements.txt
keyboard==0.13.5
openai==1.52.0
pydantic==2.9.2
streamlit==1.32.0
mysql-connector-python==8.0.27
